CONDEC-584: create three new classes to seperate ReleaseNoteRest from calculations, markdown creation, and ReleaseNote-Creation, Refactor TaskCriteriaPrioritisation to IssueMetric
