+++ src/main/java/de/uhd/ifi/se/decision/management/jira/releasenotes/ReleaseNotesCreator.java
@@ -1,0 +1,261 @@
+package de.uhd.ifi.se.decision.management.jira.releasenotes;
+
+import com.atlassian.jira.component.ComponentAccessor;
+import com.atlassian.jira.issue.Issue;
+import com.atlassian.jira.issue.IssueManager;
+import com.atlassian.jira.issue.issuetype.IssueType;
+import com.atlassian.jira.user.ApplicationUser;
+import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElement;
+import de.uhd.ifi.se.decision.management.jira.releasenotes.impl.ReleaseNoteIssueProposalImpl;
+
+import java.util.*;
+
+/**
+ * Class to compute the metrics for the proposals and compare the ratings
+ */
+public class ReleaseNotesCreator {
+	private ArrayList<ReleaseNoteIssueProposal> proposals;
+	private final List<DecisionKnowledgeElement> elementsMatchingQuery;
+	private final ReleaseNoteConfiguration config;
+	private final ApplicationUser user;
+
+
+	public ReleaseNotesCreator(List<DecisionKnowledgeElement> elementsMatchingQuery, ReleaseNoteConfiguration releaseNoteConfiguration, ApplicationUser user) {
+		this.elementsMatchingQuery = elementsMatchingQuery;
+		this.config = releaseNoteConfiguration;
+		this.user = user;
+	}
+
+	public HashMap<String, ArrayList<ReleaseNoteIssueProposal>> getMappedProposals() {
+		setMetrics();
+		compareProposals();
+		return mapProposals();
+	}
+
+
+	/**
+	 * Gather priority metrics for the Release Note Issue Proposal
+	 * <p>
+	 * sets Proposals
+	 */
+	private void setMetrics() {
+		ArrayList<ReleaseNoteIssueProposal> releaseNoteIssueProposals = new ArrayList<ReleaseNoteIssueProposal>();
+		//set up components we need to gather metrics
+		IssueManager issueManager = ComponentAccessor.getIssueManager();
+
+		//create plain array with no duplicates
+		List<String> usedKeys = new ArrayList<String>();
+		HashMap<String, Integer> reporterIssueCount = new HashMap<String, Integer>();
+		HashMap<String, Integer> resolverIssueCount = new HashMap<String, Integer>();
+		//for each DecisionKnowledgeElement create one ReleaseNoteIssueProposal element with the data
+		HashMap<String, Integer> dkLinkedCount = new HashMap<String, Integer>();
+
+		for (int i = 0; i < elementsMatchingQuery.size(); i++) {
+			DecisionKnowledgeElement dkElement = elementsMatchingQuery.get(i);
+			//add key to used keys
+			usedKeys.add(dkElement.getKey());
+			//create Release note issue proposal with the element and the count of associated decision knowledge
+			// check if DK or Comment
+			ReleaseNoteIssueProposal proposal = new ReleaseNoteIssueProposalImpl(dkElement, 0);
+			String dkKey = dkElement.getKey();
+
+			//check if it is a dk Issue or just a DK comment
+			//comments are not rated, just counted
+			if (dkKey.contains(":")) {
+				String[] parts = dkKey.split(":");
+				Integer currentCount = dkLinkedCount.get(parts[0]);
+				if (currentCount != null) {
+					currentCount += 1;
+					dkLinkedCount.put(parts[0], currentCount);
+				} else {
+					dkLinkedCount.put(parts[0], 1);
+				}
+			} else {
+				Issue issue = issueManager.getIssueByCurrentKey(dkElement.getKey());
+
+				//set priority
+				proposal.getAndSetPriority(issue);
+
+				//set count of comments
+				proposal.getAndSetCountOfComments(issue);
+
+				//set size summary
+				proposal.getAndSetSizeOfSummary();
+
+				//set size description
+				proposal.getAndSetSizeOfDescription();
+
+				//set days to complete
+				proposal.getAndSetDaysToCompletion(issue);
+
+				//set experience reporter
+				proposal.getAndSetExperienceReporter(issue, reporterIssueCount, user);
+
+				//set experience resolver
+				proposal.getAndSetExperienceResolver(issue, resolverIssueCount, user);
+
+				//add to results
+				releaseNoteIssueProposals.add(proposal);
+			}
+		}
+
+		//now check DK element links
+		for (Map.Entry<String, Integer> entry : dkLinkedCount.entrySet()) {
+			String key = entry.getKey();
+			Integer value = entry.getValue();
+			releaseNoteIssueProposals.forEach(proposal -> {
+				if (proposal.getDecisionKnowledgeElement().getKey().equals(key)) {
+					proposal.getMetrics().put(IssueMetric.COUNT_DECISION_KNOWLEDGE, value);
+				}
+			});
+		}
+		proposals = releaseNoteIssueProposals;
+	}
+
+	/**
+	 * compare all ReleaseNoteIssueProposal elements and set the rating for each category considering UserInput
+	 * compare each element with the others of the same category
+	 * scaling algorithm:
+	 * https://stackoverflow.com/questions/5294955/how-to-scale-down-a-range-of-numbers-with-a-known-min-and-max-value
+	 * alternative algorithm could be gaussian standard distribution
+	 * other alternative could be median-interval-separation
+	 */
+	private void compareProposals() {
+		EnumMap<IssueMetric, Double> userWeighting = config.getIssueMetricWeight();
+		List<IssueMetric> criteriaEnumList = IssueMetric.getOriginalList();
+
+
+		//find median
+		EnumMap<IssueMetric, Integer> medianOfProposals = RatingCalculator.getMedianOfProposals(proposals);
+
+		//for each criteria create a list of integers, so we can then compute min, max values and the scales
+		EnumMap<IssueMetric, ArrayList<Integer>> countValues = RatingCalculator.getFlatListOfValues(proposals);
+
+		//we later check in which interval the proposal would be and apply the corresponding lower and higher value
+
+		//add min and max to lists
+		//the first value of the ArrayList is for the first interval and the second is for the second interval
+		EnumMap<IssueMetric, ArrayList<Integer>> minValues = new EnumMap<>(IssueMetric.class);
+		EnumMap<IssueMetric, ArrayList<Integer>> maxValues = new EnumMap<>(IssueMetric.class);
+		RatingCalculator.getMinAndMaxValues(minValues, maxValues, countValues, medianOfProposals);
+
+		//calculate user weighting:
+		//add all and then divide by count of items
+		var ref = new Object() {
+			Double total = 0.0;
+		};
+		userWeighting.values().forEach(weight -> {
+			if (weight != null) {
+				ref.total += weight;
+			}
+		});
+		Double part = ref.total / userWeighting.values().size();
+		//then assign each the corresponding part
+		EnumMap<IssueMetric, Double> weightedValues = new EnumMap<>(IssueMetric.class);
+		userWeighting.forEach((key, value) -> {
+			if (value != null) {
+				weightedValues.put(key, value * part);
+			} else {
+				weightedValues.put(key, 0.0);
+			}
+		});
+
+
+		proposals.forEach(dkElement -> {
+			EnumMap<IssueMetric, Integer> existingCriteriaValues = dkElement.getMetrics();
+			//use ref object due to atomic problem etc.
+			var totalRef = new Object() {
+				Double total = 0.0;
+			};
+			criteriaEnumList.forEach(criteria -> {
+				double scaling = 0;
+
+				//check if criteria is in first or second interval
+				int index = 0;
+				int minVal = 1;
+				int maxVal = 5;
+				if (existingCriteriaValues.get(criteria) > medianOfProposals.get(criteria) && minValues.get(criteria).size() > 1 && maxValues.get(criteria).size() > 1) {
+					index = 1;
+					minVal = 6;
+					maxVal = 10;
+				}
+				//extra treatment for priorities, as there are no outliers and numbers are reversed
+				//we just do the scaling on all using min and max values
+				if (criteria == IssueMetric.PRIORITY) {
+					int indexMinPrio = 0;
+					int indexMaxPrio = 0;
+					if (maxValues.get(criteria).size() > 1) {
+						indexMaxPrio = 1;
+					}
+					scaling = RatingCalculator.scaleFromSmallToLarge(existingCriteriaValues.get(criteria), minValues.get(criteria).get(indexMinPrio), maxValues.get(criteria).get(indexMaxPrio), 1, 10);
+					scaling = 11 - scaling;
+				} else {
+					scaling = RatingCalculator.scaleFromSmallToLarge(existingCriteriaValues.get(criteria), minValues.get(criteria).get(index), maxValues.get(criteria).get(index), minVal, maxVal);
+				}
+				//multiply scaling with associated weighting input from user
+				double userWeight = weightedValues.get(criteria);
+				scaling *= userWeight;
+				totalRef.total += scaling;
+			});
+			//set rating
+			dkElement.setRating(Math.round(totalRef.total));
+		});
+	}
+
+
+	private HashMap<String, ArrayList<ReleaseNoteIssueProposal>> mapProposals() {
+		IssueManager issueManager = ComponentAccessor.getIssueManager();
+
+		HashMap<String, ArrayList<ReleaseNoteIssueProposal>> resultMap = new HashMap<String, ArrayList<ReleaseNoteIssueProposal>>();
+		ArrayList<ReleaseNoteIssueProposal> bugs = new ArrayList<ReleaseNoteIssueProposal>();
+		ArrayList<ReleaseNoteIssueProposal> features = new ArrayList<ReleaseNoteIssueProposal>();
+		ArrayList<ReleaseNoteIssueProposal> improvements = new ArrayList<ReleaseNoteIssueProposal>();
+		var ref = new Object() {
+			Boolean hasResult = false;
+		};
+		proposals.forEach(proposal -> {
+			Issue issue = issueManager.getIssueByCurrentKey(proposal.getDecisionKnowledgeElement().getKey());
+			IssueType issueType = issue.getIssueType();
+			Integer issueTypeId = Integer.valueOf(issueType.getId());
+			//new features
+			if (config.getFeatureMapping() != null && config.getFeatureMapping().contains(issueTypeId)) {
+				features.add(proposal);
+				ref.hasResult = true;
+			}
+			//bugs
+			//check if include bugs is false
+			if (config.getBugFixMapping() != null && config.getBugFixMapping().contains(issueTypeId) && config.getAdditionalConfiguration().get(AdditionalConfigurationOptions.INCLUDE_BUG_FIXES)) {
+				bugs.add(proposal);
+				ref.hasResult = true;
+			}
+			//improvements
+			if (config.getImprovementMapping() != null && config.getImprovementMapping().contains(issueTypeId)) {
+				improvements.add(proposal);
+				ref.hasResult = true;
+			}
+
+		});
+		if (!ref.hasResult) {
+			return null;
+		}
+		Comparator<ReleaseNoteIssueProposal> compareByRating = new Comparator<ReleaseNoteIssueProposal>() {
+			@Override
+			public int compare(ReleaseNoteIssueProposal o1, ReleaseNoteIssueProposal o2) {
+				Double rating1 = o1.getRating();
+				Double rating2 = o2.getRating();
+				return rating2.toString().compareTo(rating1.toString());
+			}
+		};
+		bugs.sort(compareByRating);
+		features.sort(compareByRating);
+		improvements.sort(compareByRating);
+
+
+		resultMap.put(ReleaseNoteCategory.BUG_FIXES.toString(), bugs);
+		resultMap.put(ReleaseNoteCategory.NEW_FEATURES.toString(), features);
+		resultMap.put(ReleaseNoteCategory.IMPROVEMENTS.toString(), improvements);
+		return resultMap;
+	}
+
+
+}
