+++ src/test/java/de/uhd/ifi/se/decision/management/jira/releasenotes/TestReleaseNoteConfiguration.java
@@ -17,7 +17,7 @@ public class TestReleaseNoteConfiguration {
 	private String endDate;
 	private String sprintId;
 	private TargetGroup targetGroup;
-	private EnumMap<TaskCriteriaPrioritisation, Double> taskCriteriaPrioritisation;
+	private EnumMap<IssueMetric, Double> issueMetric;
 	private List<Integer> bugFixMapping;
 	private List<Integer> featureMapping;
 	private List<Integer> improvementMapping;
@@ -30,7 +30,7 @@ public class TestReleaseNoteConfiguration {
 		endDate = "2020-01-12";
 		sprintId = "15";
 		targetGroup = TargetGroup.DEVELOPER;
-		taskCriteriaPrioritisation = TaskCriteriaPrioritisation.toDoubleEnumMap();
+		issueMetric = IssueMetric.toDoubleEnumMap();
 		bugFixMapping = new ArrayList<Integer>();
 		featureMapping = new ArrayList<Integer>();
 		improvementMapping = new ArrayList<Integer>();
@@ -69,9 +69,9 @@ public class TestReleaseNoteConfiguration {
 
 
 	@Test
-	public void testTaskCriteriaPrioritisation() {
-		config.setTaskCriteriaPrioritisation(taskCriteriaPrioritisation);
-		assertEquals(taskCriteriaPrioritisation, config.getTaskCriteriaPrioritisation());
+	public void testissueMetric() {
+		config.setIssueMetricWeight(issueMetric);
+		assertEquals(issueMetric, config.getIssueMetricWeight());
 	}
 
 	@Test
